pyproject.toml
[project]
name = "xstitchlab"
version = "0.1.0"
description = "Cross-stitch pattern generation toolchain: Image → Pixelation → Color mapping → Pattern output"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "colormath>=3.0.0",
    "openai>=1.0.0",
    "fpdf2>=2.7.0",
    "typer>=0.9.0",
    "streamlit>=1.28.0",
    "rich>=13.0.0",
    "streamlit-image-coordinates>=0.4.0",
]

[project.scripts]
xstitch = "xstitchlab.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["xstitchlab"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
